==> Rcode/4_stats_birthweight.R <==
# -----------------------------------------------------------
# Author: A.J. Paijmans
#
# Script Name: 4_stats_birthweight
#
# Purpose: This script is used to run the models for pup birth weight.
#
# Date: 2024-03-20
# -----------------------------------------------------------


library(here)
library(readxl)
library(tidyverse)
library(lme4)
library(lmerTest) 
library(DHARMa)
library(sjPlot)
library(cowplot)



#~~~~~~~~~~~~~~~~#
#  Load data  ####
#~~~~~~~~~~~~~~~~#

pup_data <- openxlsx::read.xlsx(here("Data", "Processed", "filtered_pup_survival.xlsx"), detectDates = T) %>%
  mutate(Pup_Sex = as.factor(Pup_Sex)) %>%
  mutate(Year = as.factor(Year)) %>%
  mutate(Survival = as.factor(Survival))

str(pup_data)



#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
#  Full model: effect of pup/mum sMLH on pup birth mass  ####
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#

## ---- m1_BM --------
#~~ Birth mass model incl maternal effects
m1birthmass <- lm(Pup_BirthWeight ~ sMLH_msat39_pup
                  + Pup_Sex
                  + Year
                  + sMLH_msat39_mum
                  + Mum_Age,
                  # + (1 | uniqueID_mum), # including mother ID as a random effect did not change the results significantly and the model was a poorer fit
                  # In addition, adding mother ID as random effect may make it problematic as sMLH mum would be a confounding factor.
                  data = pup_data)

summary(m1birthmass) 

## ---- m1_BM_assumptions --------

#~~ Model assumptions
testDispersion(m1birthmass) # tests for over/underdispersion
plotQQunif(m1birthmass) # detect overall deviations from the expected distribution. Is data normally distributed?
plotResiduals(m1birthmass) # plot residuals against the predicted value
# Outlier test is significant, although the QQplot looks good

#~~ Further investigation of the outliers
simulationOutput <- simulateResiduals(fittedModel = m1birthmass) 
testOutliers(simulationOutput)
# There are 7 outliers out of 342 obs. The frequency of outliers is significantly higher than expected (2.04% against 0.79% expected)
# According to the vignette, a certain number of outliers are expected 'at random'. It could also hint at overdispersion, but the test for dispersion
# showed no indication for that. The DHARMa vignette also mentions that the visual indicators of the residuals are a lot more informative than the p-values.
# Taken together with the fact that all plots look good, it is not likely to be a problem.

##---- chunk_end

#~~ Save model so that we can load it for the figure
saveRDS(m1birthmass, file = here("Data", "Processed", "m1_birthmass.rds"))



#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
#  Model 2: effect of pup sMLH on pup birth mass, excluding mums  ####
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#

## ---- m2_BM --------
#~~ Birth mass model excl maternal effects
m2birthmass <- lm(Pup_BirthWeight ~ sMLH_msat39_pup
                  + Pup_Sex
                  + Year,
                  data = pup_data)

summary(m2birthmass)

## ---- m2_BM_assumptions --------

#~~ Model assumptions
testDispersion(m2birthmass)
plotQQunif(m2birthmass)
plotResiduals(m2birthmass)

# Outlier test is significant, although the QQplot looks good
#~~ Further investigation of the outliers
simulationOutput <- simulateResiduals(fittedModel = m2birthmass) 
testOutliers(simulationOutput)
# There are 16 outliers out of 885 obs. 
# Also here, the test for dispersion shows no indication for overdispersion, and the QQplot looks good.
# Therefore, the model fit is likely to be OK.

##---- chunk_end

#~~ Save model so that we can load it for the figure
saveRDS(m2birthmass, file = here("Data", "Processed", "m2_birthmass.rds"))



#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
#  Stats table for both models  ####
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#

## ---- BM_table --------

# Statistical table showing parameter estimates for both models
# Labels
tab_label <- c(
  `(Intercept)` = "Intercept",
  sMLH_msat39_pup = "pup sMLH",
  Pup_SexM = "pup sex [M]",
  sMLH_msat39_mum = "mother sMLH",
  Mum_Age = "mother age",
  Year2018 = "season [2019]",
  Year2019 = "season [2020]",
  Year2020 = "season [2021]")

# Table
# print so it saves but doesn't show the html table, 
# which doesn't display nicely in the pdf generated by Rmarkdown
print(tab_model(m1birthmass, m2birthmass,
                pred.labels = tab_label,
                title = "Pup birth mass",
                dv.labels = c("(a) model incl. maternal effect",
                              "(b) model excl. maternal effect"),
                show.stat=T, 
                string.stat = "t value",
                file = here("Tables", "Table_BW_full_model_vs_no_mat_NEW.html")))

# Make a screenshot of saved htlm table and save as a png
# so that it can be shown in Rmarkdown pdf
webshot::webshot(here("Tables", "Table_BW_full_model_vs_no_mat_NEW.html"), 
                 file=here("Tables", "Table_BW_full_model_vs_no_mat_NEW.png"), delay=2, vheight = 450, vwidth = 700)

##---- chunk_end
